requirements.txt
flask==2.2.3
PyPDF2==3.0.1
Werkzeug==2.2.3
google-generativeai==0.3.1
Flask-Session==0.5.0
